itest: move test `testDisconnectingTargetPeer`
Hence finish an old TODO.
